Rename test_cpp_transporter_traveler_type.h -> test_cpp_transporter_traveler_types.h
